restructure deps
