+++ .github/workflows/ci.yml
@@ -1,8 +1,9 @@
 name: CI
 
 on:
-  push:
   pull_request:
+    branches:
+      - master
 
 jobs:
   build:
