+++ libs/cat/dashboard/src/lib/dashboard.component.html
@@ -1,2 +1,1 @@
 <p>Dashboardos!</p>
-<noda-editor (connectionChange)="configConnectionChanged($event)" [source]="nodes"></noda-editor>
