+++ libs/cat/dashboard/src/lib/dashboard.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { NodaNodeData, NodaConnectionChange } from '@ratcat/noda';
 
 @Component({
 	selector: 'cat-dashboard',
@@ -8,27 +7,6 @@ import { NodaNodeData, NodaConnectionChange } from '@ratcat/noda';
 })
 export class DashboardComponent {
 
-	nodes: NodaNodeData[] = [
-		{
-			id: 1,
-			name: 'Sandbox',
-
-		},
-		{
-			id: 2,
-			name: 'Production',
-		},
-		{
-			id: 3,
-			name: 'Default',
-		}
-	];
-
 	constructor() {
 	}
-
-	configConnectionChanged($event: NodaConnectionChange) {
-		console.log($event);
-	}
-
 }
