+++ libs/cat/dashboard/src/lib/dashboard.module.ts
@@ -1,7 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { NodaModule } from '@ratcat/noda';
 import { DashboardComponent } from './dashboard.component';
 
 @NgModule({
@@ -14,8 +13,7 @@ import { DashboardComponent } from './dashboard.component';
 				pathMatch: 'full',
 				component: DashboardComponent
 			}
-		]),
-		NodaModule
+		])
 	]
 })
 export class DashboardModule {}
