+++ libs/domain/.eslintrc.json
@@ -1,0 +1,33 @@
+{
+	"extends": [
+		"../../.eslintrc.json"
+	],
+	"ignorePatterns": [
+		"!**/*"
+	],
+	"overrides": [
+		{
+			"files": [
+				"*.ts",
+				"*.tsx",
+				"*.js",
+				"*.jsx"
+			],
+			"rules": {}
+		},
+		{
+			"files": [
+				"*.ts",
+				"*.tsx"
+			],
+			"rules": {}
+		},
+		{
+			"files": [
+				"*.js",
+				"*.jsx"
+			],
+			"rules": {}
+		}
+	]
+}
