+++ libs/domain/README.md
@@ -1,0 +1,7 @@
+# domain
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test domain` to execute the unit tests via [Jest](https://jestjs.io).
