+++ libs/domain/project.json
@@ -1,0 +1,29 @@
+{
+	"root": "libs/domain",
+	"sourceRoot": "libs/domain/src",
+	"projectType": "library",
+	"targets": {
+		"lint": {
+			"executor": "@nrwl/linter:eslint",
+			"outputs": [
+				"{options.outputFile}"
+			],
+			"options": {
+				"lintFilePatterns": [
+					"libs/domain/**/*.ts"
+				]
+			}
+		},
+		"test": {
+			"executor": "@nrwl/jest:jest",
+			"outputs": [
+				"coverage/libs/domain"
+			],
+			"options": {
+				"jestConfig": "libs/domain/jest.config.js",
+				"passWithNoTests": true
+			}
+		}
+	},
+	"tags": []
+}
