+++ libs/domain/src/index.ts
@@ -1,1 +1,2 @@
+export * from './lib/domain';
 export * from './lib/cat-env';
