+++ libs/domain/src/lib/domain.ts
@@ -1,0 +1,3 @@
+export function domain(): string {
+	return 'domain';
+}
