+++ libs/domain/tsconfig.lib.json
@@ -1,0 +1,14 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"outDir": "../../dist/out-tsc",
+		"declaration": true,
+		"types": []
+	},
+	"include": [
+		"**/*.ts"
+	],
+	"exclude": [
+		"**/*.spec.ts"
+	]
+}
