+++ libs/shared/error/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/error-reporter';
+export * from './lib/errors';
