+++ libs/shared/utils/README.md
@@ -1,0 +1,7 @@
+# shared-utils
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-utils` to execute the unit tests via [Jest](https://jestjs.io).
