+++ libs/shared/utils/project.json
@@ -1,0 +1,29 @@
+{
+	"root": "libs/shared/utils",
+	"sourceRoot": "libs/shared/utils/src",
+	"projectType": "library",
+	"targets": {
+		"lint": {
+			"executor": "@nrwl/linter:eslint",
+			"outputs": [
+				"{options.outputFile}"
+			],
+			"options": {
+				"lintFilePatterns": [
+					"libs/shared/utils/**/*.ts"
+				]
+			}
+		},
+		"test": {
+			"executor": "@nrwl/jest:jest",
+			"outputs": [
+				"coverage/libs/shared/utils"
+			],
+			"options": {
+				"jestConfig": "libs/shared/utils/jest.config.js",
+				"passWithNoTests": true
+			}
+		}
+	},
+	"tags": []
+}
