+++ libs/shared/utils/tsconfig.json
@@ -5,6 +5,9 @@
 	"references": [
 		{
 			"path": "./tsconfig.lib.json"
+		},
+		{
+			"path": "./tsconfig.spec.json"
 		}
 	],
 	"compilerOptions": {
