+++ libs/shared/utils/tsconfig.spec.json
@@ -1,0 +1,22 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"outDir": "../../../dist/out-tsc",
+		"module": "commonjs",
+		"types": [
+			"jest",
+			"node"
+		]
+	},
+	"include": [
+		"**/*.test.ts",
+		"**/*.spec.ts",
+		"**/*.test.tsx",
+		"**/*.spec.tsx",
+		"**/*.test.js",
+		"**/*.spec.js",
+		"**/*.test.jsx",
+		"**/*.spec.jsx",
+		"**/*.d.ts"
+	]
+}
